+++ src/main/java/com/arangodb/internal/net/AccessType.java
@@ -1,7 +1,7 @@
 /*
  * DISCLAIMER
  *
- * Copyright 2016 ArangoDB GmbH, Cologne, Germany
+ * Copyright 2018 ArangoDB GmbH, Cologne, Germany
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,25 +20,12 @@
 
 package com.arangodb.internal.net;
 
-import java.io.IOException;
-
 /**
  * @author Mark Vollmary
  *
  */
-public interface HostHandler {
+public enum AccessType {
 
-	Host get(HostHandle hostHandle, AccessType accessType);
+	WRITE, READ, DIRTY_READ
 
-	void success();
-
-	void fail();
-
-	void reset();
-
-	void confirm();
-
-	void close() throws IOException;
-
-	void closeCurrentOnError();
 }